JUDDI-701 8 more issues fixed
git-svn-id: https://svn.apache.org/repos/asf/juddi/trunk@1540387 13f79535-47bb-0310-9956-ffa450edef68

+++ juddi-core/src/main/java/org/apache/juddi/api/impl/UDDISubscriptionImpl.java
@@ -750,7 +750,7 @@ public class UDDISubscriptionImpl extends AuthenticatedService implements UDDISu
 				catch(ConfigurationException ce) { 
 					throw new FatalErrorException(new ErrorMessage("errors.configuration.Retrieval"));
 				}
-				newChunkToken.setExpiresAfter(new Date(System.currentTimeMillis() + chunkExpirationMinutes * 60 * 1000));
+				newChunkToken.setExpiresAfter(new Date(System.currentTimeMillis() + ((long)chunkExpirationMinutes * 60L * 1000L)));
 				
 				em.persist(newChunkToken);
